feat: pulled pieces out into own files. made useWidth hook. changed mobile header to still use same header image as desktop.
