+++ src/pages/home/BannerSection.tsx
@@ -1,0 +1,22 @@
+const BannerSection = () => {
+  return (
+    <>
+      <div className={`banner-section`}>
+        <div className="col-lg-6 col-sm-8">
+          <div className="banner-text">
+            <h5 className="text-white">
+              We're building the multi-chain crypto bridge so you can swap
+              between major crypto networks.
+            </h5>
+            <p className="text-green">
+              <strong>Invest in the project today.</strong>
+            </p>
+            <p>Earn up to 500% APR in our reward hub.</p>
+          </div>
+        </div>
+      </div>
+    </>
+  );
+};
+
+export default BannerSection;
